+++ R/fastNaiveBayes.gaussian.default.R
@@ -1,7 +1,7 @@
 #' @export
 #' @import Matrix
 #' @rdname fastNaiveBayes.gaussian
-fastNaiveBayes.gaussian.default <- function(x, y, std_threshold = 0.01, sparse = FALSE, ...) {
+fastNaiveBayes.gaussian.default <- function(x, y, sparse = FALSE, ...) {
   if (class(x)[1] != "dgCMatrix") {
     if (!is.matrix(x)) {
       x <- as.matrix(x)
@@ -17,29 +17,32 @@ fastNaiveBayes.gaussian.default <- function(x, y, std_threshold = 0.01, sparse =
     stop("X and Y must be equal length")
   }
 
-  probability_table <- lapply(levels(y), function(level) {
-    x_level <- x[y == level, ]
-    if (ncol(x) == 1) {
-      x_level <- as.matrix(x_level)
-    }
+  n <- tabulate(y)
+  if (sparse) {
+    probability_table <- lapply(levels(y), function(level) {
+      x_level <- x[y == level, ]
+      if (ncol(x) == 1) {
+        x_level <- as.matrix(x_level)
+      }
 
-    if (sparse) {
+
       means <- Matrix::colMeans(x_level, na.rm = TRUE)
       mat_means <- matrix(means, nrow = nrow(x_level), ncol = ncol(x_level), byrow = TRUE)
       stddev <- sqrt(Matrix::colSums((x_level - mat_means)^2) / (nrow(x_level) - 1))
-    } else {
-      means <- base::colMeans(x_level, na.rm = TRUE)
-      mat_means <- matrix(means, nrow = nrow(x_level), ncol = ncol(x_level), byrow = TRUE)
-      stddev <- sqrt(colSums((x_level - mat_means)^2) / (nrow(x_level) - 1))
-    }
+      return(list(level = level, means = means, stddev = stddev))
+    })
+  }else{
+    rs <- rowsum(x,y)
+    means <- rs/n
+    stddev <- sqrt((rowsum(x^2,y)-2*means*rs+n*means^2)/(n-1))
+    probability_table <- lapply(levels(y), function(level){
+      return(list(level = level, means = means[level,],
+                  stddev = stddev[level,]))
+    })
+  }
 
-    # Apply laplace
-    stddev[stddev == 0] <- std_threshold
-
-    return(list(level = level, means = means, stddev = stddev))
-  })
-
-  priors <- table(y) / nrow(x)
+  names(n) <- levels(y)
+  priors <- n/nrow(x)
   structure(list(
     probability_table = probability_table,
     priors = priors,